requirements.txt
djangorestframework==3.16.0
python-dotenv==1.1.1
black==25.1.0
flake8==7.3.0
isort==6.0.1
pylint==4.0.2
pylint-django==2.6.1
